Fix version checking in rails helper
The way the version checking has worked before is it was using bitmovin <6 JS API `bitdash` if the version number included "5". This was probably intended as a major version check, but unfortunately it also catches latest versions like 7.5.

This commit fixes that by explicitly checking the major part of the version. It also safeguards against versions like "Latest Version 5".